requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
statsmodels==0.14.2
prophet==1.1.5
matplotlib==3.9.0
seaborn==0.13.2
pyyaml==6.0.2
joblib==1.4.2
